requirements.txt
streamlit
groq
python-dotenv
PyPDF2
docx
extract-docx
colorama
